// src/app/pages/reports/filemanager/model/model.ts
export class FolderFile {
    id: string;
    denumire: string;
    parinteId: string;
    description: string;
    isFile:number = 0;
    nameFile:string;
}

// export interface updateFolder {
//     label: string;
//     data: string;
//     expandedIcon: string;
//     collapsedIcon: string;
//     children: string;
// }